+++ src/renderer/widgets/Staking/BondNominate/default/ui/BondForm.tsx
@@ -104,7 +104,7 @@ const Signatories = () => {
     });
   }, [signatories, balances, network]);
 
-  if (!network) {
+  if (!network || signatoriesWithBalance.length < 2) {
     return null;
   }
