+++ src/renderer/widgets/Staking/Restake/default/ui/RestakeForm.tsx
@@ -92,7 +92,7 @@ const Signatories = () => {
     });
   }, [signatories, balances, network]);
 
-  if (!network) {
+  if (!network || signatoriesWithBalance.length < 2) {
     return null;
   }
