+++ src/renderer/widgets/Staking/Unstake/index.ts
@@ -1,4 +1,4 @@
-export { Unstake } from './default/components/Unstake';
+export { Unstake } from './default/ui/Unstake';
 export { unstakeModel } from './default/model/unstake-model';
 export { unstakeModel as unstakeModelShards } from './shards/model/unstake-model';
 export { Unstake as UnstakeShards } from './shards/components/Unstake';
